+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+[Added]
+- Added support for Czech language for reporting error 
+
 [Changed]
 - Removed all deprecated items from dictionaries
+- Filling email is now required for reporting error
 
 ## [v1.0.0]
 [Added]